+++ src/Enums.cs
@@ -1,0 +1,19 @@
+namespace TFrengler.CSSelenium
+{
+    public enum Browser
+    {
+        EDGE        = 1,
+        FIREFOX     = 2,
+        CHROME      = 3
+    }
+
+    public enum Platform
+    {
+        WINDOWS, LINUX
+    }
+
+    public enum Architecture
+    {
+        x64, x86
+    }
+}
